feat: Add post route and update HTML response
This commit adds a new post route '/submit' to the application in our index.ts file. It also optimizes the homepage route by storing the index.html content in a variable. In addition, the application's index.html is updated with a form, changing the main view from a 'Hello World!' greeting to a user-input form.